Renaming Instrument.py to _Instrument.py s.t. the module is distinguishable from its class in the python module path.
