+++ LabExT/Instruments/InstrumentAPI/_Instrument.py
